+++ src/utils.cc
@@ -1,14 +1,12 @@
 #include "utils.h"
 
-#include <Eigen/src/Core/util/Meta.h>
-
 double Utils::L1Distance(const Point &pt1, const Point &pt2) {
     if (pt1.size() != pt2.size()) {
         spdlog::error("Vectors must be of the same size");
     }
 
-    Eigen::Map<const Eigen::VectorXf> v1(pt1.data(), static_cast<Eigen::Index>(pt1.size()));
-    Eigen::Map<const Eigen::VectorXf> v2(pt2.data(), static_cast<Eigen::Index>(pt2.size()));
+    Eigen::Map<const Eigen::VectorXd> v1(pt1.data(), static_cast<Eigen::Index>(pt1.size()));
+    Eigen::Map<const Eigen::VectorXd> v2(pt2.data(), static_cast<Eigen::Index>(pt2.size()));
 
     return static_cast<double>((v1 - v2).lpNorm<1>());
 }
